put props in proper places to access hardcoded info in timeline and timelineevent
